ebay_client/requirements.txt
pyzmq==15.4.0
ebaysdk==2.1.4
